Adds contract/output snapshot tests
